[BUGFIX beta] Move public modules out of the `-private` folder
Closes #3998
