+++ addon/-private/serializers.js
@@ -2,9 +2,9 @@
   @module ember-data
 */
 
-import JSONAPISerializer from "ember-data/-private/serializers/json-api-serializer";
-import JSONSerializer from "ember-data/-private/serializers/json-serializer";
-import RESTSerializer from "ember-data/-private/serializers/rest-serializer";
+import JSONAPISerializer from "ember-data/serializers/json-api";
+import JSONSerializer from "ember-data/serializers/json";
+import RESTSerializer from "ember-data/serializers/rest";
 
 export {
   JSONAPISerializer,
