+++ addon/-private/system/model.js
@@ -3,7 +3,7 @@
 */
 
 import Model from "ember-data/-private/system/model/model";
-import attr from "ember-data/-private/system/model/attributes";
+import attr from "ember-data/attr";
 import RootState from "ember-data/-private/system/model/states";
 import Errors from "ember-data/-private/system/model/errors";
