--- addon/-private/system/relationships.js
@@ -1,13 +1,0 @@
-/**
-  @module ember-data
-*/
-
-import belongsTo from "./relationships/belongs-to";
-import hasMany from "./relationships/has-many";
-
-import "ember-data/-private/system/relationships/ext";
-
-export {
-  belongsTo,
-  hasMany
-};
