+++ addon/-private/system/relationships/belongs-to.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
 import { assert, warn } from "ember-data/-private/debug";
-import Model from 'ember-data/-private/system/model';
+import Model from 'ember-data/model';
 import normalizeModelName from "ember-data/-private/system/normalize-model-name";
 
 /**
