+++ addon/-private/transforms.js
@@ -1,4 +1,4 @@
-import Transform from "ember-data/-private/transforms/base";
+import Transform from "ember-data/transform";
 import NumberTransform from "ember-data/-private/transforms/number";
 import DateTransform from "ember-data/-private/transforms/date";
 import StringTransform from "ember-data/-private/transforms/string";
