+++ addon/-private/transforms/date.js
@@ -20,7 +20,7 @@ import Ember from 'ember';
   @extends DS.Transform
   @namespace DS
  */
-import Transform from "ember-data/-private/transforms/base";
+import Transform from "ember-data/transform";
 
 export default Transform.extend({
   deserialize(serialized) {
