+++ addon/relationships.js
@@ -1,1 +1,14 @@
-export * from 'ember-data/-private/system/relationships';
+/**
+  @module ember-data
+*/
+
+import belongsTo from "ember-data/-private/system/relationships/belongs-to";
+import hasMany from "ember-data/-private/system/relationships/has-many";
+
+import "ember-data/-private/system/relationships/ext";
+
+export {
+  belongsTo,
+  hasMany
+};
+
