+++ addon/store.js
@@ -1,1 +1,3 @@
-export { default } from "ember-data/-private/system/store";
+import Store from "ember-data/-private/system/store";
+
+export default Store;
